python-backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
PyJWT>=2.8.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
pymongo>=4.6.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
Werkzeug>=3.0.0
